requirements.txt
gunicorn
asgiref
django-environ
django==4
djangorestframework==3.13.1
drf-yasg==1.20.0
celery
channels
django_redis
channels_redis
redis
requests
xmltodict